backend/requirements.txt
Django==5.0.6
h5py==3.11.0
matplotlib==3.8.4
numpy==2.4.0
pandas==2.3.3
Pillow==12.0.0
scikit_learn==1.8.0
tensorflow==2.20.0
